+++ src/hittable_list.h
@@ -22,7 +22,7 @@ class hittable_list : public hittable {
         bool hit(const ray& r, interval ray_t, hit_record& rec) const override {
             hit_record temp_rec;
             bool hit_anything = false;
-            double closest = ray_t.max;
+            float closest = ray_t.max;
             
             for (const auto& object : objects) {
                 if (object->hit(r, interval(ray_t.min, closest), temp_rec)) {
