+++ src/interval.h
@@ -3,25 +3,25 @@
 
 class interval {
     public:
-        double min, max;
+        float min, max;
 
         interval() : min(+infinity), max(-infinity) {}
 
-        interval(double min, double max) : min(min), max(max) {}
+        interval(float min, float max) : min(min), max(max) {}
 
-        double size() const {
+        float size() const {
             return max - min;
         }
 
-        bool contains(double x) const {
+        bool contains(float x) const {
             return x >= min && x <= max;
         }
 
-        bool surrounds(double x) const {
+        bool surrounds(float x) const {
             return x > min && x < max;
         }
 
-        double clamp(double x) const {
+        float clamp(float x) const {
             if (x < min) return min;
             if (x > max) return max;
             return x;
